LayoutTests/fast/events/ios/meaningful-click-when-focusing-body.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<head>
<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
<style>
body, html {
    font-size: 18px;
    font-family: system-ui;
    width: 100%;
    height: 100%;
    margin: 0;
}
</style>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<script>
jsTestIsAsync = true;

addEventListener("load", async () => {
    description("This test exercises the 'meaningful click' heuristic in an editable body element, and requires WebKitTestRunner.");

    didHandleTap = false;
    didHandleTapAsMeaningfulClick = false;
    testRunner.installDidHandleTapCallback(wasMeaningful => {
        didHandleTap = true;
        didHandleTapAsMeaningfulClick = wasMeaningful;
    });

    await UIHelper.activateElementAndWaitForInputSession(document.body);
    await UIHelper.waitForDoubleTapDelay();
    testPassed("Focused body element");
    shouldBeTrue("didHandleTap");
    shouldBeTrue("didHandleTapAsMeaningfulClick");

    testRunner.clearTestRunnerCallbacks();
    didHandleTap = false;
    didHandleTapAsMeaningfulClick = false;
    testRunner.installDidHandleTapCallback(wasMeaningful => {
        didHandleTap = true;
        didHandleTapAsMeaningfulClick = wasMeaningful;
    });

    await UIHelper.activateAt(100, 100);
    await UIHelper.waitForDoubleTapDelay();
    testPassed("Tapped inside body element");
    shouldBeTrue("didHandleTapAsMeaningfulClick");

    testRunner.clearTestRunnerCallbacks();
    document.getElementById("text").remove();
    finishJSTest();
});
</script>
</head>
<body contenteditable>
    <p id="text">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aenean laoreet nisi felis, eu scelerisque dolor imperdiet in. Sed at porttitor purus, in placerat justo. Mauris at augue ac purus viverra molestie. Fusce ultrices ante at pellentesque lobortis. Sed pellentesque, ipsum eu semper mollis, arcu lectus dignissim enim, in maximus tortor ligula id ipsum. Nulla dignissim diam a mi tempor porta. Etiam tempor ac augue non dapibus. Donec euismod vel turpis in ultrices. Ut accumsan ultrices arcu vitae venenatis. In urna velit, accumsan sit amet arcu vel, vestibulum pulvinar leo. Aliquam vulputate euismod ultrices. Sed congue vestibulum libero quis ornare.</p>
    <div id="description"></div>
    <div id="console"></div>
</body>
</html>
